# Conflicts:
#	collect-client/pom.xml
#	collect-core/src/main/java/org/openforis/collect/io/SurveyRestoreJob.java
#	collect-core/src/main/java/org/openforis/collect/manager/SurveyManager.java
#	collect-core/src/main/java/org/openforis/collect/manager/UserGroupManager.java
#	collect-core/src/main/java/org/openforis/collect/model/CollectSurvey.java
#	collect-core/src/main/java/org/openforis/collect/model/SurveySummary.java
#	collect-core/src/main/java/org/openforis/collect/model/UserGroup.java
#	collect-core/src/main/java/org/openforis/collect/persistence/SurveyDao.java
#	collect-core/unused/InstitutionManager.java
#	collect-server/src/main/java/org/openforis/collect/designer/viewmodel/BaseVM.java
#	collect-server/src/main/java/org/openforis/collect/designer/viewmodel/NewSurveyParametersPopUpVM.java
#	collect-server/src/main/java/org/openforis/collect/designer/viewmodel/SurveyImportVM.java
#	collect-server/src/main/java/org/openforis/collect/designer/viewmodel/SurveySelectVM.java
#	collect-server/src/main/java/org/openforis/collect/web/controller/SurveyController.java
#	collect-server/src/main/java/org/openforis/collect/web/controller/UserGroupController.java
#	collect-server/src/main/resources/org/openforis/collect/application-context-server.xml
#	collect-server/unused/ClientInstitutionManager.java
#	collect-web/collect-webapp/src/main/webapp/WEB-INF/labels/designer/label.properties
#	collect-web/collect-webapp/src/main/webapp/WEB-INF/view/zul/designer/component/new_survey_parameters_popup.zul
#	collect-web/collect-webapp/src/main/webapp/WEB-INF/view/zul/designer/survey_import_popup.zul
#	collect-web/collect-webapp/src/main/webapp/WEB-INF/view/zul/designer/survey_select.zul

+++ collect-core/src/main/java/org/openforis/collect/metamodel/SimpleSurveyCreationParameters.java
@@ -20,7 +20,7 @@ public class SimpleSurveyCreationParameters implements Serializable {
 	private List<ListItem> values = new ArrayList<ListItem>();
 	private List<String> imagery = new ArrayList<String>();
 	private SamplingPointGenerationSettings samplingPointGenerationSettings;
-	private long userGroupId;
+	private int userGroupId;
 	
 	public String getName() {
 		return name;
@@ -62,11 +62,11 @@ public class SimpleSurveyCreationParameters implements Serializable {
 		this.samplingPointGenerationSettings = samplingPointGenerationSettings;
 	}
 	
-	public long getUserGroupId() {
+	public int getUserGroupId() {
 		return userGroupId;
 	}
 	
-	public void setUserGroupId(long userGroupId) {
+	public void setUserGroupId(int userGroupId) {
 		this.userGroupId = userGroupId;
 	}
